+++ src/components/Button/index.ts
@@ -1,0 +1,2 @@
+export { default as Button } from './component'
+export { default as AsyncButton } from './AsyncButton'
